requirements.txt
requests==2.31.0
bs4==0.0.2
markdownify==0.12.1
tqdm==4.66.2
flask==3.0.3
fonttools>=4.0.0
Pillow>=9.0.0
ddddocr>=1.4.0